requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
Flask==3.0.2
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.1
google-generativeai==0.4.0
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.62.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Markdown==3.5.2
MarkupSafe==2.1.5
numpy==1.26.4
openai==1.13.3
pathlib==1.0.1
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
pydub==0.25.1
PyPDF2==3.0.1
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
scipy==1.12.0
sniffio==1.3.1
sounddevice==0.4.6
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
youtube-transcript-api==0.6.2
